pyproject.toml
[project]
name = "uwotm8"
dynamic = ["version"]
description = "Converting American English to British English and removing AI-generated text patterns"
authors = [{name = "Chris Dean", email = "chris@digitaltactics.co.uk"}]
repository = "https://github.com/DigitalTactics/uwotm8"
documentation = "https://digitaltactics.github.io/uwotm8/"
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"
packages = [
  {include = "uwotm8"}
]

[project.scripts]
uwotm8 = "uwotm8.convert:main"

[project.optional-dependencies]
docx = ["python-docx>=0.8.11"]

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry]
version = "0.0.0"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
breame = "^0.1.2"
python-docx = {version = ">=0.8.11", optional = true}

[tool.poetry.extras]
docx = ["python-docx"]

[tool.poetry.group.dev.dependencies]
pytest = "8.3.4"
pytest-cov = "6.0.0"
deptry = "^0.16.2"
mypy = "1.15.0"
pre-commit = "4.1.0"
tox = "4.24.1"
ruff = "^0.9.6"
mkdocs = "1.6.1"

[tool.poetry.group.docs.dependencies]
mkdocs = "1.6.1"
mkdocs-material = "9.6.3"
mkdocstrings = {extras = ["python"], version = "^0.26.1"}
mkdocs-mermaid2-plugin = "^1.2.1"
markdown-mdantic = "^2.1.1"
pygments = "^2.19.1"
griffe-typingdoc = "^0.2.7"
mkdocs-git-revision-date-localized-plugin = "^1.3.0"
mkdocs-glightbox = "^0.4.0"
mkdocs-minify-plugin = "^0.8.0"

[tool.poetry.requires-plugins]
poetry-dynamic-versioning = { version = ">=1.0.0,<2.0.0", extras = ["plugin"] }

[tool.poetry-dynamic-versioning]
enable = true

[tool.mypy]
files = ["uwotm8"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true
lint.select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
lint.ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["uwotm8"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
